more work integrating audio and video headers ...

+++ src/buffers/VideoBuffer.cpp
@@ -76,7 +76,7 @@ unsigned int VideoBuffer::getMaxSize(){
 }
 
 
-int VideoBuffer::getFps(){
+float VideoBuffer::getFps(){
     return fps;
 }
