+++ src/buffers/VideoBuffer.h
@@ -52,7 +52,7 @@ public:
 	unsigned int size();                            // total size of the buffer
 	unsigned int getMaxSize();                         // max size of the buffer
 
-	int getFps();                                   // fps of the video source
+	float getFps();                                   // fps of the video source
 
     Timestamp getLastTimestamp();                 // timestamp for the last frame
 	TimeDiff getTotalTime();                      // total time of the buffer in microseconds
